Smart Home Automation using IoT
A smart home automation system is based on ensuring security and making user life easier. It contains a large number of sensors which can control or monitor objects distributed in three-dimensional space. The sensors can be specialized in measuring temperature, humidity, pressure, light, noise, dust air, and so on. This is the Arduino IDE code for all the sensors used in this project.